refactor: rename slashtags2 -> slashtags
