+++ src/components/BlocksWidget.tsx
@@ -8,7 +8,7 @@ import BaseFeedWidget from './BaseFeedWidget';
 import { TFeedWidget } from '../store/types/widgets';
 import { useSlashfeed } from '../hooks/widgets';
 import { decodeWidgetFieldValue, SUPPORTED_FEED_TYPES } from '../utils/widgets';
-import { useSlashtags2 } from '../hooks/slashtags2';
+import { useSlashtags } from '../hooks/slashtags';
 import { __E2E__ } from '../constants/env';
 
 const mapping = {
@@ -41,7 +41,7 @@ const BlocksWidget = ({
 	onPressIn?: () => void;
 }): ReactElement => {
 	const { t } = useTranslation('slashtags');
-	const { webRelayClient, webRelayUrl } = useSlashtags2();
+	const { webRelayClient, webRelayUrl } = useSlashtags();
 	const { config, fields, loading } = useSlashfeed({ url });
 	const [data, setData] = useState(fields);
