+++ src/components/ContactImage.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react';
 import { StyleProp, ViewStyle } from 'react-native';
-import { useProfile2 } from '../hooks/slashtags2';
+import { useProfile } from '../hooks/slashtags';
 import ProfileImage from './ProfileImage';
 
 const ContactImage = ({
@@ -12,7 +12,7 @@ const ContactImage = ({
 	size?: number;
 	style?: StyleProp<ViewStyle>;
 }): ReactElement => {
-	const { profile } = useProfile2(url);
+	const { profile } = useProfile(url);
 	return (
 		<ProfileImage style={style} url={url} image={profile.image} size={size} />
 	);
