+++ src/components/PriceChart.tsx
@@ -24,7 +24,7 @@ import { Pair } from '@synonymdev/slashtags-widget-price-feed/types/lib/reader';
 import { IThemeColors } from '../styles/themes';
 import { SlashFeedJSON, TGraphPeriod } from '../store/types/widgets';
 import useColors from '../hooks/colors';
-import { useSlashtags2 } from '../hooks/slashtags2';
+import { useSlashtags } from '../hooks/slashtags';
 
 export type Change = {
 	color: keyof IThemeColors;
@@ -159,7 +159,7 @@ const PriceChart = ({
 	period: TGraphPeriod;
 	style?: StyleProp<ViewStyle>;
 }): ReactElement => {
-	const { webRelayClient, webRelayUrl } = useSlashtags2();
+	const { webRelayClient, webRelayUrl } = useSlashtags();
 	const [history, setHistory] = useState<THistory>();
 
 	useEffect(() => {
