+++ src/hooks/widgets.ts
@@ -4,7 +4,7 @@ import b4a from 'b4a';
 
 import { SlashFeedJSON } from '../store/types/widgets';
 import { SUPPORTED_FEED_TYPES, decodeWidgetFieldValue } from '../utils/widgets';
-import { useSlashtags2 } from './slashtags2';
+import { useSlashtags } from './slashtags';
 
 type Field = {
 	name: string;
@@ -33,7 +33,7 @@ export const useSlashfeed = (options: {
 	loading: boolean;
 	failed: boolean;
 } => {
-	const { webRelayClient, webRelayUrl } = useSlashtags2();
+	const { webRelayClient, webRelayUrl } = useSlashtags();
 	const [config, setConfig] = useState<SlashFeedJSON>(
 		cache[options.url]?.config,
 	);
